refactor: move SettingsStore into core/bess (no backend-specific dependencies)
SettingsStore had zero backend/FastAPI dependencies -- only stdlib
(errno, json, logging, os, tempfile) -- so its home in backend/ was
organizational, not technical. core/bess/ is the documented domain layer;
backend/ is meant to depend on it, never the reverse.

This relocation is prep work, not the fix itself: core/bess/ha_api_controller.py
currently can't hold a live SettingsStore reference (backend -> core is the
only allowed dependency direction), which is why ha_controller.sensors has
to be a manually-synced cache instead of a live view (see #332). Moving the
module is the low-risk half of that change; turning controller.sensors into
a computed property is a separate follow-up given its wider blast radius
(constructor changes across every inverter-controller subclass).

- git mv backend/settings_store.py -> core/bess/settings_store.py
- Updated the only two module-level imports (backend/app.py, backend/api.py)
  and their equivalents in test files; all other references were already
  through a bess_controller.settings_store attribute and needed no change
- Dockerfile: removed settings_store.py from the flat backend-file COPY list
  since it's now included via the existing `COPY core/ /app/core/`
