+++ backend/api.py
@@ -33,11 +33,11 @@ from api_dataclasses import (
 )
 from fastapi import APIRouter, HTTPException, Query
 from loguru import logger
-from settings_store import VALID_PLATFORMS
 
 from core.bess import time_utils
 from core.bess.health_check import describe_failing_checks, run_system_health_checks
 from core.bess.savings_aggregator import DEFAULT_COUNTS, build_buckets
+from core.bess.settings_store import VALID_PLATFORMS
 from core.bess.time_utils import get_period_count
 
 router = APIRouter()
